+++ Assets/Scripts/Player/PlayerStats.cs
@@ -6,7 +6,7 @@ using Assets.Scripts.EnumTypes;
 public class PlayerStats : MonoBehaviour
 {
     public float playerMaxHP = 9;
-    public int playerDamage = 1;
+    public float playerDamage = 1f;
     public int playerMilk = 2;
     public float playerMovementSpeed = 150f;
     public float playerMaxStamina = 50f;
